Revert "Use memcpy, etc. from libc crate n xmlparse.rs"
This reverts commit 0531566619641e73c501ab9b42782f90daf3baa3.

+++ src/expat_config_h.rs
@@ -1,2 +1,2 @@
-
-pub const XML_CONTEXT_BYTES: isize = 1024;
+use libc::c_int;
+pub const XML_CONTEXT_BYTES: c_int = 1024;
